Remove ExprSequence
Now that we can break the old API, we don't need an `ExprSequence`
anymore. This commit removes `ExprSequence` and replaces all of its uses
with `Sequence`. Note that I had to temporarily make `sequence.h` public
because we currently include it in a "public" header because we still
generate the code for `Expr::mkConst<Sequence>()`.