[FP] Fix overly restrictive assertion
Fixes https://github.com/cvc5/cvc5-projects/issues/327. When generating
refinement lemmas, we were asserting that the concrete value of
conversion from real to a floating-point number was either normal or
subnormal. However, it can also be zero. This commit fixes the assertion
and adds a modified version of the regression that triggers the same
error and terminates quickly with the fix.